ADM-402 move sourceControl store to config folder
