+++ frontend/src/store.ts
@@ -3,7 +3,7 @@ import stepperReducer from './context/stepper/StepperSlice'
 import configReducer from './context/config/configSlice'
 import jiraVerifyResponseReducer from './context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 import pipelineToolResponseReducer from './context/config/pipelineTool/pipelineToolVerifyResponse/pipelineToolVerifyResponseSlice'
-import sourceControlVerifyResponseReducer from './context/sourceControl/sourceControlVerifyResponse/sourceControlVerifyResponseSlice'
+import sourceControlVerifyResponseReducer from './context/config/sourceControl/sourceControlVerifyResponse/sourceControlVerifyResponseSlice'
 import saveMetricsSettingReducer from './context/Metrics/metricsSlice'
 
 export const store = configureStore({
